+++ Ignition.Core/Ignition.Core.csproj
@@ -164,7 +164,7 @@
     <Compile Include="HtmlHelpers\DynamicPlaceholder.cs" />
     <Compile Include="HtmlHelpers\GetDynamicKeyAllowedRenderings.cs" />
     <Compile Include="HtmlHelpers\GetDynamicPlaceholderChromeData.cs" />
-    <Compile Include="Extensions\GlassExtensions.cs" />
+    <Compile Include="HtmlHelpers\GlassHelper.cs" />
     <Compile Include="Models\BaseModels\IModelBase.cs" />
     <Compile Include="Models\BaseModels\IModelBaseWithMetadata.cs" />
     <Compile Include="Models\BaseModels\INeedsChildren.cs" />
